Adding VM queue lenght data
